requirements.txt
langchain-groq
langgraph
duckduckgo-search==6.3.3
streamlit
python-dotenv
pytest
